Split out RoomListDBusInterface
